# apps/api/src/scienceresearch/flask_app.py
"""Flask composition root for the Vue-hosted offline workbench.

The legacy ``Handler`` remains the behaviour-preserving HTTP adapter while
the product is migrated route-by-route.  This adapter runs it in-process as
WSGI; it does not open a second HTTP port in production.
"""

from __future__ import annotations

from email.message import Message
from io import BytesIO
from pathlib import Path
from flask import Flask, Response, jsonify, redirect, request, send_from_directory

from .bootstrap import create_workbench
from .domain.errors import AppError
from .web import Handler


class _CapturedLegacyHandler(Handler):
    """Minimal BaseHTTPRequestHandler surface used by the retained adapter."""

    _sessions = Handler._sessions

    def __init__(self, *, app: object, method: str, path: str, headers: Message, body: bytes) -> None:
        self.app = app  # type: ignore[assignment]
        self.command = method
        self.path = path
        self.headers = headers
        self.rfile = BytesIO(body)
        self.wfile = BytesIO()
        self.status_code = 500
        self.response_headers: list[tuple[str, str]] = []

    def send_response(self, code: int, message: str | None = None) -> None:
        del message
        self.status_code = code

    def send_header(self, keyword: str, value: str) -> None:
        self.response_headers.append((keyword, value))

    def end_headers(self) -> None:
        return None


def _request_headers() -> Message:
    headers = Message()
    for key, value in request.headers.items():
        headers[key] = value
    return headers


def _legacy_path(path: str) -> str:
    query = request.query_string.decode("latin-1")
    return "/" + path + (f"?{query}" if query else "")


def _rewrite_legacy_html(body: bytes) -> bytes:
    """Keep legacy forms and links inside the compatibility frame.

    The response is generated by trusted local code.  Rewriting only absolute
    same-origin attributes prevents an old form submission from replacing the
    Vue host while retaining the legacy no-JavaScript behaviour.
    """
    text = body.decode("utf-8")
    for attribute in ("href", "action", "src"):
        text = text.replace(f"{attribute}='/", f"{attribute}='/_legacy/")
        text = text.replace(f'{attribute}="/', f'{attribute}="/_legacy/')
    return text.encode("utf-8")


def _rewrite_location(value: str) -> str:
    if value.startswith("/") and not value.startswith("/_legacy/"):
        return "/_legacy" + value
    return value


def _legacy_response(workbench: object, path: str, *, rewrite_html: bool) -> Response:
    handler = _CapturedLegacyHandler(
        app=workbench,
        method=request.method,
        path=_legacy_path(path),
        headers=_request_headers(),
        body=request.get_data(cache=True),
    )
    if request.method == "GET":
        handler.do_GET()
    elif request.method == "POST":
        handler.do_POST()
    else:
        return Response("Method not allowed", status=405, content_type="text/plain; charset=utf-8")

    body = handler.wfile.getvalue()
    content_type = next((value for key, value in handler.response_headers if key.lower() == "content-type"), "application/octet-stream")
    if rewrite_html and content_type.startswith("text/html"):
        body = _rewrite_legacy_html(body)

    response = Response(body, status=handler.status_code, content_type=content_type)
    for key, value in handler.response_headers:
        lowered = key.lower()
        if lowered in {"content-type", "content-length"}:
            continue
        response.headers.add(key, _rewrite_location(value) if lowered == "location" and rewrite_html else value)
    return response


def _frontend_dist() -> Path:
    return Path(__file__).resolve().parents[4] / "apps" / "web" / "dist"


def create_app(*, data_dir: Path | None = None, frontend_dist: Path | None = None) -> Flask:
    """Create the production Flask application without a global database state."""
    project_root = Path(__file__).resolve().parents[4]
    runtime_dir = (data_dir or project_root / "storage" / "runtime").resolve()
    dist_dir = (frontend_dist or _frontend_dist()).resolve()
    app = Flask(__name__)
    app.config.update(SECRET_KEY=None, JSON_AS_ASCII=False)
    app.extensions["scienceresearch_workbench"] = create_workbench(runtime_dir)
    app.extensions["scienceresearch_frontend_dist"] = dist_dir

    @app.errorhandler(AppError)
    def application_error(error: AppError) -> Response:
        return jsonify({"code": error.code, "detail": error.message}), error.status

    @app.get("/healthz")
    def healthz() -> Response:
        return jsonify({"status": "ok", "service": "scienceresearch"})

    @app.get("/api/v1/catalog")
    def catalog() -> Response:
        return jsonify(app.extensions["scienceresearch_workbench"].catalog())

    @app.route("/api/v1/contexts", methods=["GET", "POST"])
    def contexts() -> Response:
        workbench = app.extensions["scienceresearch_workbench"]
        if request.method == "GET":
            return jsonify({"contexts": workbench.list_contexts()})
        payload = request.get_json(silent=True)
        if not isinstance(payload, dict):
            return jsonify({"code": "INVALID_INPUT", "detail": "JSON object is required"}), 400
        created = workbench.create_context(
            str(payload.get("type", "topic")),
            str(payload.get("name", "")),
            str(payload.get("problem", "")),
            str(payload.get("goal", "")),
            str(payload.get("owner", "author")),
        )
        return jsonify(created), 201

    @app.get("/api/v1/contexts/<context_id>/workflows")
    def context_workflows(context_id: str) -> Response:
        workbench = app.extensions["scienceresearch_workbench"]
        workflows = workbench.workflows_for_context(context_id)
        for workflow in workflows:
            workflow["achievement_cards"] = workbench.achievement_cards(context_id, workflow["id"])
        return jsonify({"workflows": workflows, "progress": workbench.context_progress(context_id)})

    @app.get("/_legacy/")
    def legacy_trailing_slash() -> Response:
        """Prevent the SPA fallback from consuming a historical legacy URL."""
        return redirect("/_legacy", code=308)

    @app.route("/_legacy", defaults={"path": ""}, methods=["GET", "POST"])
    @app.route("/_legacy/<path:path>", methods=["GET", "POST"])
    def legacy(path: str) -> Response:
        return _legacy_response(app.extensions["scienceresearch_workbench"], path, rewrite_html=True)

    @app.route("/api", defaults={"path": "api"}, methods=["GET", "POST"])
    @app.route("/api/<path:path>", methods=["GET", "POST"])
    def api(path: str) -> Response:
        legacy_path = path if path == "api" else f"api/{path}"
        return _legacy_response(app.extensions["scienceresearch_workbench"], legacy_path, rewrite_html=False)

    @app.get("/assets/<path:asset_path>")
    def frontend_asset(asset_path: str) -> Response:
        if not dist_dir.is_dir():
            return Response("Vue production assets are not built. Run npm.cmd run build.", status=503, content_type="text/plain; charset=utf-8")
        return send_from_directory(dist_dir / "assets", asset_path)

    @app.get("/")
    @app.get("/<path:frontend_path>")
    def frontend(frontend_path: str = "") -> Response:
        del frontend_path
        index = dist_dir / "index.html"
        if not index.is_file():
            return Response("Vue production assets are not built. Run npm.cmd run build.", status=503, content_type="text/plain; charset=utf-8")
        return send_from_directory(dist_dir, "index.html")

    return app