# packages/contracts/scienceresearch.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScienceResearch API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/contexts": {
      "get": {
        "tags": [
          "contexts",
          "workflows",
          "achievement cards"
        ],
        "summary": "List Contexts",
        "operationId": "list_contexts_api_v1_contexts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextListDTO"
                }
              }
            }
          },
          "503": {
            "description": "Context database dependency is unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "contexts",
          "workflows",
          "achievement cards"
        ],
        "summary": "Create Context",
        "operationId": "create_context_api_v1_contexts_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedContextDTO"
                }
              }
            }
          },
          "400": {
            "description": "Request body is not a JSON object",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "Context input is invalid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          },
          "503": {
            "description": "Context database dependency is unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contexts/{context_id}": {
      "get": {
        "tags": [
          "contexts",
          "workflows",
          "achievement cards"
        ],
        "summary": "Context Detail",
        "operationId": "context_detail_api_v1_contexts__context_id__get",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextDTO"
                }
              }
            }
          },
          "404": {
            "description": "Context does not exist",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          },
          "503": {
            "description": "Context database dependency is unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contexts/{context_id}/workflows": {
      "get": {
        "tags": [
          "contexts",
          "workflows",
          "achievement cards"
        ],
        "summary": "Context Workflows",
        "operationId": "context_workflows_api_v1_contexts__context_id__workflows_get",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListDTO"
                }
              }
            }
          },
          "404": {
            "description": "Context does not exist",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          },
          "503": {
            "description": "Workflow database dependency is unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "row_version": {
            "type": "integer",
            "title": "Row Version"
          },
          "lifecycle_state": {
            "type": "string",
            "title": "Lifecycle State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "problem",
          "goal",
          "owner",
          "created_at",
          "row_version",
          "lifecycle_state"
        ],
        "title": "ContextDTO",
        "description": "Context row shape returned by the legacy list endpoint."
      },
...
      "ContextListDTO": {
        "properties": {
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/ContextDTO"
            },
            "type": "array",
            "title": "Contexts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contexts"
        ],
        "title": "ContextListDTO"
      },
      "CreatedContextDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "problem": {
            "type": "string",
            "title": "Problem"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "problem",
          "goal",
          "owner",
          "created_at"
        ],
        "title": "CreatedContextDTO",
        "description": "Exact fields returned by the legacy context create command."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JsonValue": {},
      "ProgressDTO": {
        "properties": {
          "context_id": {
            "type": "string",
            "title": "Context Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "expected_total": {
            "type": "integer",
            "title": "Expected Total"
          },
          "snapshot_incomplete": {
            "type": "boolean",
            "title": "Snapshot Incomplete"
          },
          "warnings": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Warnings"
          },
          "missing_work_package_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Work Package Ids"
          },
          "extra_work_package_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extra Work Package Ids"
          },
          "unassigned_work_package_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unassigned Work Package Ids"
          },
          "areas": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Areas"
          },
          "workflows": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Workflows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context_id",
          "total",
          "completed",
          "ratio",
          "expected_total",
          "snapshot_incomplete",
          "warnings",
          "missing_work_package_ids",
          "extra_work_package_ids",
          "unassigned_work_package_ids",
          "areas",
          "workflows"
        ],
        "title": "ProgressDTO",
        "description": "Legacy progress projection; warning variants intentionally remain JSON-typed."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "context_id": {
            "type": "string",
            "title": "Context Id"
          },
          "work_package_id": {
            "type": "string",
            "title": "Work Package Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "completion_row_version": {
            "type": "integer",
            "title": "Completion Row Version"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "completed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed By"
          },
          "area_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          },
          "selection_status": {
            "type": "string",
            "title": "Selection Status"
          },
          "selection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Reason"
          },
          "author_confirmed": {
            "type": "boolean",
            "title": "Author Confirmed"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "row_version": {
            "type": "integer",
            "title": "Row Version"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          },
          "inputs_json": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Inputs Json"
          },
          "achievement_cards": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object",
              "minProperties": 0
            },
            "type": "array",
            "title": "Achievement Cards"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "context_id",
          "work_package_id",
          "status",
          "created_at",
          "is_completed",
          "completion_row_version",
          "completed_at",
          "completed_by",
          "area_id",
          "template_id",
          "skill_id",
          "selection_status",
          "selection_reason",
          "author_confirmed",
          "mode",
          "row_version",
          "archived",
          "inputs_json",
          "achievement_cards"
        ],
        "title": "WorkflowDTO",
        "description": "Complete workflow row plus the legacy achievement card projection."
      },
      "WorkflowListDTO": {
        "properties": {
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowDTO"
            },
            "type": "array",
            "title": "Workflows"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressDTO"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflows",
          "progress"
        ],
        "title": "WorkflowListDTO"
      },
...
      "APIErrorDTO": {
        "additionalProperties": false,
        "description": "Stable problem payload used by every target API error response.",
        "properties": {
          "code": {
            "examples": [
              "INVALID_INPUT"
            ],
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "examples": [
              "请求参数无效"
            ],
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "requestId": {
            "examples": [
              "5f0d5b3a6cbf4f21916d1bd5f47b28c1"
            ],
            "minLength": 1,
            "pattern": "^[0-9a-f]{32}$",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail",
          "requestId"
        ],
        "title": "APIErrorDTO",
        "type": "object"
      }
    }
  }
}